research/airflow-research/pyproject.toml
[project]
name = "airflow-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "airflow-dbt-python>=3.1.0",
    "airflow-provider-duckdb>=0.2.0",
    "apache-airflow==2.10.5",
    "dbt-bigquery>=1.10.1",
    "dbt-core>=1.10.9",
    "duckdb>=1.3.2",
]
